+++ packages/dart_pdf_editor_flutter_gpu/lib/src/backend_stats.dart
@@ -39,6 +39,7 @@ class FlutterGpuTileBackendStats {
   int analyticAtlasFallbacks = 0;
   int analyticSparseAtlasSkips = 0;
   int analyticTextFallbackRuns = 0;
+  int glyphBindingReuses = 0;
   int clipPathsCompiled = 0;
   int clipMaskRebuilds = 0;
   int paperClearTiles = 0;
@@ -125,6 +126,7 @@ class FlutterGpuTileBackendStats {
         'analyticAtlasFallbacks': analyticAtlasFallbacks,
         'analyticSparseAtlasSkips': analyticSparseAtlasSkips,
         'analyticTextFallbackRuns': analyticTextFallbackRuns,
+        'glyphBindingReuses': glyphBindingReuses,
         'clipPathsCompiled': clipPathsCompiled,
         'clipMaskRebuilds': clipMaskRebuilds,
         'paperClearTiles': paperClearTiles,
@@ -210,6 +212,7 @@ class FlutterGpuTileBackendStats {
     analyticAtlasFallbacks = 0;
     analyticSparseAtlasSkips = 0;
     analyticTextFallbackRuns = 0;
+    glyphBindingReuses = 0;
     clipPathsCompiled = 0;
     clipMaskRebuilds = 0;
     paperClearTiles = 0;
@@ -275,6 +278,7 @@ class FlutterGpuTileBackendStats {
       'analyticAtlasFallbacks=$analyticAtlasFallbacks '
       'analyticSparseSkips=$analyticSparseAtlasSkips '
       'analyticFallbackRuns=$analyticTextFallbackRuns '
+      'glyphBindingReuses=$glyphBindingReuses '
       'clips=$clipPathsCompiled clipRebuilds=$clipMaskRebuilds '
       'paperClearTiles=$paperClearTiles '
       'paperOnlyTiles=$paperOnlyTiles '